426 ext

+++ plugin/trino-memory/src/main/java/io/trino/plugin/memory/MemorySplitManager.java
@@ -14,6 +14,7 @@
 package io.trino.plugin.memory;
 
 import com.google.common.collect.ImmutableList;
+import com.google.inject.Inject;
 import io.trino.spi.connector.ConnectorSession;
 import io.trino.spi.connector.ConnectorSplit;
 import io.trino.spi.connector.ConnectorSplitManager;
@@ -23,8 +24,6 @@ import io.trino.spi.connector.ConnectorTransactionHandle;
 import io.trino.spi.connector.Constraint;
 import io.trino.spi.connector.DynamicFilter;
 import io.trino.spi.connector.FixedSplitSource;
-
-import javax.inject.Inject;
 
 import java.util.List;
 import java.util.OptionalLong;
@@ -60,7 +59,7 @@ public final class MemorySplitManager
 
         List<MemoryDataFragment> dataFragments = metadata.getDataFragments(table.getId());
 
-        int totalRows = 0;
+        long totalRows = 0;
 
         ImmutableList.Builder<ConnectorSplit> splits = ImmutableList.builder();
